Switch from port ranges to use PSIDs.
This adds a side table mapping IPv4 addresses to their PSID parameters.
Given those parameters we can compute a PSID from a port on an incoming
packet.  The IPv4 address and the PSID together form the key into the
binding table.  We still need to switch to use the PodHashMap for the
binding table itself.