Fix bad bimodal distribution in PodHashMap performance self-test
In the PodHashMap insertion test, don't create the values via an
ffi.new() with a complex initializer.  Besides not being how we would do
this in practice, this was causing the test loop to be blacklisted,
invalidating later results.  The result would be tests that ran really
slowly.  I had seen this in the past and chalked up the bad bimodal
distribution to something random in the memory subsystem (that needed
investigating) but didn't realize it was actually something random in
LuaJIT.  I didn't always see it because there is something random about
the way that LuaJIT decides to start tracing.  (I really wish it were
deterministic!)