Define data types for packed binding table
When a packet comes in, we will extract its IPv4 address and its port.
Even if the packet is IPv6, we can extract this information from the
payload.  Thanks to Marcel Wiget for this idea.  We will then look up
the A and K parameters for the IPv4 address from a table, and use those
parameters to compute the PSID from the port.  This AK table will be
implemented as a sorted vector, partitioning the IPv4 address space into
regions with the given AK values.  We'll compile an optimized binary
search for this table, using the dynasm techniques from the PodHashMap.
We will also have a lookaside table of IPv6 addresses for the lwAFTR.