Add snabb-lwaftr subcommand to run without physical hardware
This adds a new "snabb-lwaftr run-nohw" subcommand, which allows running the
AFTR without the need for physical NICs. It uses a raw socket in each end of
the lwAFTR. This means that it is possible to create a pair of dummy network
interfaces and have Snabb attach to them:

  modprobe dummy numdummies=2
  ip -6 addr add $AFTR_V6_ADDR dev dummy0  # For the B4 side
  ip link set dummy0 address $AFTR_V6_MAC
  ip addr add $AFTR_V4_ADDR dev dummy1     # For the Internet side
  ip link set dummy1 $AFTR_V4_MAC
  snabb-lwaftr run-nohw -b bindingtable.lua -c config.lua -B dummy0 -I dummy1

Of course, additional setup on top of the above commands is needed to have a
complete testing setup: the packets arriving to dummy1 need to be sent to the
Internet. That is possible using iptables, by bridging dummy1 with a NIC that
reaches to Internet, or by passing the NIC which reaches Internet as argument
to the "-I" flag. Also, a B4 must be configured to put packets into dummy0:
again, by using iptables, by bridging dummy0, or any other sutiable means.
Make sure to update the addresses in the AFTR configuration file to reflect
your setup.

Alternatively, if using qemu for testing, one can make it use a TUN/TAP
device in the host side, for example if the B4 is runs as a virtual machine,
it is possible to pass "-netdev tap,ifname=tun-aftr -net nic" to qemu, and
then passing "-B tun-aftr" when executing snabb-lwaftr. This method skips the
dummy interface and connects the AFTR directly to the host-side TUN/TAP
interface that qemu creates. Note that the IPv6 address and AFTR-B4-side MAC
addresses in the AFTR configuration file correspond to those of the TUN/TAP
device. Again, edit the addresses in your AFTR configuration file to match.