PSID info map in binding table covers only managed addresses
This changes the RangeMap to have a "default value" which it inserts
whereever there is a gap between entries.  This will allow the user to
distinguish between entries that are present and those that are absent.

From the perspective of the binding table, this lets us know if an IPv4
address is managed by us.  It's a speedup.  It could cause the creation
of more entries in the address map; this problem is mitigated by the
extent to which the provisioned set of IPv4 addresses is contiguous.

Change the lwaftr to use this new facility to determine when to
hairpin.