vita: explicitly configure and route encrypted trafffic via SPI
This “costs” an extra mandatory configuration value (the
Security Parameter Index). On the other hand, we can now test any number
of routes on a single node. Plus, its in the spirit of RFC 4301.

This also fixes a bug related to the handling of IPv4 TTL, where the
inner header had its TTL not decremented.