KeyManager: change default sa_ttl from one week to one day
Generally speaking, because of PFS, the more often we renegotiate our
keys, the less damaging a compromised Vita node will be with respect to
traffic that has been recorded prior to the compromise. So there is a
valid argument to reduce the default sa_ttl even further, e.g. to
hours or even minutes.

On the other hand, key cycling might induce loss of packets that are
in-flight when the keys are changed. WireGuard solves this by keeping
multiple “versions” of each SA around, but Vita has no such mechanism as
of yet.

For now I think cycling keys twice a day (remember: keys are
renegotiated once half the sa_ttl has expired) is a good general
trade-off between steady packet flow and key hygiene, which can be
further adjusted by users depending on their needs.

  “Fragen über Fragen, warum ich neue Nummer hole’. Wechselt ihr nicht
   jeden Tag eure Unterhose?”